(fix) webpack fix snackbar
