--- npm-debug.log
@@ -1,45 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ '/usr/local/bin/node', '/usr/local/bin/npm', 'run', 'mocha' ]
-2 info using npm@3.7.3
-3 info using node@v5.8.0
-4 verbose run-script [ 'premocha', 'mocha', 'postmocha' ]
-5 info lifecycle good-times@0.1.0~premocha: good-times@0.1.0
-6 silly lifecycle good-times@0.1.0~premocha: no script for premocha, continuing
-7 info lifecycle good-times@0.1.0~mocha: good-times@0.1.0
-8 verbose lifecycle good-times@0.1.0~mocha: unsafe-perm in lifecycle true
-9 verbose lifecycle good-times@0.1.0~mocha: PATH: /usr/local/lib/node_modules/npm/bin/node-gyp-bin:/Users/poseidon/Documents/groupProjects/good-times/node_modules/.bin:/usr/local/bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/Applications/Postgres.app/Contents/Versions/latest/bin
-10 verbose lifecycle good-times@0.1.0~mocha: CWD: /Users/poseidon/Documents/groupProjects/good-times
-11 silly lifecycle good-times@0.1.0~mocha: Args: [ '-c', 'mocha --recursive --compilers js:babel-core/register' ]
-12 silly lifecycle good-times@0.1.0~mocha: Returned: code: 4  signal: null
-13 info lifecycle good-times@0.1.0~mocha: Failed to exec mocha script
-14 verbose stack Error: good-times@0.1.0 mocha: `mocha --recursive --compilers js:babel-core/register`
-14 verbose stack Exit status 4
-14 verbose stack     at EventEmitter.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/lifecycle.js:239:16)
-14 verbose stack     at emitTwo (events.js:100:13)
-14 verbose stack     at EventEmitter.emit (events.js:185:7)
-14 verbose stack     at ChildProcess.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/spawn.js:24:14)
-14 verbose stack     at emitTwo (events.js:100:13)
-14 verbose stack     at ChildProcess.emit (events.js:185:7)
-14 verbose stack     at maybeClose (internal/child_process.js:850:16)
-14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:215:5)
-15 verbose pkgid good-times@0.1.0
-16 verbose cwd /Users/poseidon/Documents/groupProjects/good-times
-17 error Darwin 15.4.0
-18 error argv "/usr/local/bin/node" "/usr/local/bin/npm" "run" "mocha"
-19 error node v5.8.0
-20 error npm  v3.7.3
-21 error code ELIFECYCLE
-22 error good-times@0.1.0 mocha: `mocha --recursive --compilers js:babel-core/register`
-22 error Exit status 4
-23 error Failed at the good-times@0.1.0 mocha script 'mocha --recursive --compilers js:babel-core/register'.
-23 error Make sure you have the latest version of node.js and npm installed.
-23 error If you do, this is most likely a problem with the good-times package,
-23 error not with npm itself.
-23 error Tell the author that this fails on your system:
-23 error     mocha --recursive --compilers js:babel-core/register
-23 error You can get information on how to open an issue for this project with:
-23 error     npm bugs good-times
-23 error Or if that isn't available, you can get their info via:
-23 error     npm owner ls good-times
-23 error There is likely additional logging output above.
-24 verbose exit [ 1, true ]
