--- src/components/ActivityList.js
@@ -1,32 +1,0 @@
-// import React from 'react';
-// import { connect } from 'react-redux';
-// import { bindActionCreators } from 'redux';
-// import * as actionCreators from '../actions';
-
-// export class ActivitiesList extends React.Component {
-//   render() {
-//     return (
-//       <div>
-//         <h3>{this.props.title}</h3>
-//         <div>{this.props.children}</div>
-//       </div>
-//     )
-//   }
-// }
-
-// // ActivitiesList.propTypes = {
-// //   children: React.PropTypes.node,
-// //   title: PropTypes.string.isRequired
-// // }
-
-// const mapStateToProps = (state) => ({
-// });
-
-// const mapDispatchToProps = (dispatch) => ({
-//   actions: bindActionCreators(actionCreators, dispatch)
-// });
-
-// export default connect(
-//   mapStateToProps,
-//   mapDispatchToProps
-// )(ActivitiesList);
