+++ src/components/AuthenticatedComponent.js
@@ -4,43 +4,43 @@ import {push} from 'redux-router';
 
 export function requireAuthentication(Component) {
 
-    class AuthenticatedComponent extends React.Component {
+  class AuthenticatedComponent extends React.Component {
 
-        componentWillMount () {
-            this.checkAuth(this.props.isAuthenticated);
-        }
-
-        componentWillReceiveProps (nextProps) {
-            this.checkAuth(nextProps.isAuthenticated);
-        }
-
-        checkAuth (isAuthenticated) {
-            if (!isAuthenticated) {
-                let redirectAfterLogin = this.props.location.pathname;
-                this.props
-                    .dispatch(push(`/login?next=${redirectAfterLogin}`));
-            }
-        }
-
-        render () {
-            return (
-                <div>
-                    {this.props.isAuthenticated === true
-                        ? <Component {...this.props}/>
-                        : null
-                    }
-                </div>
-            )
-
-        }
+    componentWillMount () {
+      this.checkAuth(this.props.isAuthenticated);
     }
 
-    const mapStateToProps = (state) => ({
-        token: state.auth.token,
-        userName: state.auth.userName,
-        isAuthenticated: state.auth.isAuthenticated
-    });
+    componentWillReceiveProps (nextProps) {
+      this.checkAuth(nextProps.isAuthenticated);
+    }
 
-    return connect(mapStateToProps)(AuthenticatedComponent);
+    checkAuth (isAuthenticated) {
+      if (!isAuthenticated) {
+        let redirectAfterLogin = this.props.location.pathname;
+        this.props
+          .dispatch(push(`/login?next=${redirectAfterLogin}`));
+      }
+    }
+
+    render () {
+      return (
+        <div>
+          {this.props.isAuthenticated === true
+            ? <Component {...this.props}/>
+            : null
+          }
+        </div>
+      )
+
+    }
+  }
+
+  const mapStateToProps = (state) => ({
+    token: state.auth.token,
+    userName: state.auth.userName,
+    isAuthenticated: state.auth.isAuthenticated
+  });
+
+  return connect(mapStateToProps)(AuthenticatedComponent);
 
 }
