+++ src/containers/index.js
@@ -1,7 +1,7 @@
 export ActivitiesContainer from './ActivitiesContainer';
 export App from './App';
 export Filter from './Filter';
-export PlanBuilder from './PlanBuilder';
+export PlanBuilderContainer from './PlanBuilderContainer';
 export Root from './Root';
 export UpdateProfile from './UpdateProfile';
 export UserProfile from './UserProfile';
